Move some modules to Internal submodule
