+++ lib/DBus/Internal/Types.hs
@@ -19,7 +19,7 @@
 -- You should have received a copy of the GNU General Public License
 -- along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-module DBus.Types where
+module DBus.Internal.Types where
 
 import           Control.Monad (liftM, when, (>=>))
 import           Control.Exception (Exception, handle, throwIO)
